Add a voxel plot to test_plot
