+++ tests/test_plot/plots.xml
@@ -23,4 +23,10 @@
     <background>0 0 0</background>
   </plot>
 
+  <plot id="4" type="voxel">
+    <pixels>100 100 10</pixels>
+    <origin>0. 0. 0.</origin>
+    <width>20 20 10</width>
+  </plot>
+
 </plots>
