+++ tests/test_plot/results_true.dat
@@ -1,1 +1,1 @@
-f76183da6a581c17f9a1572e1f1573d784e3f2e3efa3e7a3bd8786991df7a086e0b5f7f1c6b00343fa61b601ec70ad619b5ff5d8120af32559b882ca57be6768
+01ecda0f3820a49c8a41d8dc47d1e5c58767a04301621c2437231fcc04401ddea47b67d0529ca56a32d4d97b4f1416a2e0b6120d3bdc87d74a7e9889758a8808
